There is weird behavior with the left alt modifier in the us layout
Added a new test case to show in a simple file what is happening. I've
seen this stops happening while addding or removing seemingly unrelated
components to the keymap. In the full processed us layout this does not
happen, but as soon as we start remove things left alt stops working.

I need to take a closer look at this.