Fix vanishing modifiers when compatibility was empty
In xkb_parser_simplify_layout() I was doing an early return if there is
no compatibility section, but we can't do this because we haven't
finished translating actions from our parser representation into the
internal representation.

I also renamed xkb_parser_resolve_compatibility() into
xkb_parser_simplify_layout() because the old name sort of implies that
it is only necessary if there is a compatibility section to resolve, but
this is always a mandatory step after parsing.