Fix bug in real modifier mapping computation
I was trying to be too clever. I can't update curr_unmapped_element
while I'm assuming it will always be the currently chosen element to be
mapped.

I opted by symplifying the algorithm so it's easier to understand an
verify its correctness. We now look for the next unmapped element at the
beginning of the loop. I could probably just store the element being
mapped in another variable, but I didn't like how complex the algorithm
was starting to look. If we ever care about the performance of this we
could look into this idea. I'm not even sure the complex algorithm is
faster than the new one.

I further simplified the test case that was able to reproduce the bug.
Even though this test does not check a special feature, I think the more
test cases we have the better. Which is why I left it here.