Add more tests about conflicts between interpret and symbol
This checks what happens to the virtual modifier encoding when we have
it being defined both from an interpret statement and in the key
statement of the symbols section.

Looks like the virtual modifier in the symbols section will override the
one in the interpret statement, whenever the key statement contains a
vmods or actions statement.