I can't map perfect hash to num_mod_keys elements
The previous implementation of modifier_equality_test() was mapping the
results of the bit mask perfect hash to a num_mod_keys sized array, this
is wrong as the perfect hash most likely will produce values outside of
this range.

I had to either make it a 32 element array, or use a 32 element int
array to create the lookup table that resolves to the actual bit
position of the bit mask of which we take the perfect hash. I used the
2nd approach, just because I had to choose one, and I wanted to
implement the function that creates this lookup table. It may be slower
in the end, I don't know.