Make keys that latch mods be modifier keys too, just don't test them
We need them to be in the modifier key linked list as these their
complement is expected to be the set of non modifier keys.

We now add them to the list, but skip them when setting the modifier
state. This has the problem that they still count towards the maximum
modifier limit when they actually shouldn't but meh, I've never seen
latching modifier keys in the keyboard database, they seem rather
uncommon.