Add an echo mode to pymk
The idea of this mode is to be able to debug targets without executing
them. It easily transforms all ex() calls so they just print the
resolved command.

I was accomplishing something similar before by prepending 'echo' to the
command, the problem is this doesn't work if the ex call is inside
another function.

There is a  problem with this mode for commands that use its output to
drive the control flow. For example, an if statement that uses the
return value of a call to ex. These commands will return None, and may
cause issues. How can we improve this?...