Fix several bugs in the testing code
The binary search was completely broken, I had the end condition
flipped. The maximum for pressed_keys is 2^num_mod_keys not
num_mod_keys^2.

Fixed the style of the output, there were some printfs left in code that
should be concatenating into a string. If all stages of a test are
successful, then we don't show OK multiple times, just once for the
keymap name.

Checked that all tests run on valgrind with 0 errors.