Skip writing modifier mappings that have 'none' as modifiers for level>1
In the parser, we don't support having the emprty modifier mask mapped
to any level different than the 1st one. We do this because even though
libxkbcommon parses this, I haven't taken the time to figure out what is
the resulting behavior of doing so.

Even then, if a level had a virtual modifier, and this virtual modifier
was resolved to 'none', then we would get a modifier mapping with
modifiers equal to 'none'. If we get these mappings in types, then we
skip them in the writer.

I'm doing this to fix a difference we were having with the KEYPAD type
where it had the Level1 in the input keymap but the one in Level2 in the
writer's output. I'm not sure what was the root cause, looks like the
Level2 having 'none' as modifier made the default keysyms have the
content of Level2?. Keymaps passed the symbol equality test but not the
modifier equality test (in the first run).