In the modifier equality test, also compare when no modifiers are pressed
I just discovered the symbol equality test can pass, and still have 2
keymaps that have different symbols when no modifier is pressed. This
will happen if somehow a type is being interpreted in different ways.
Compare keymaps for this case too in the modifier equality test.

Also improved the error message of the modifier equality test.