Fix LED testing, we don't want them to be identical
Comparing LEDs by index was a mistake, but on top of that we don't
really want to test LEDs to be exactly the same because there are kinds
of indicators in the original specification tha we don't support.
instead. we only check indicators that can be activated by a combination
of modifier keys. If the original keyboar has unactivatable indicators
then it will have more indicators defined than the one of our output.

Also added another initial test that checks that parsing will not crash,
we use fork() to isolate these tests. This was useful to fix a null
pointer dereference that happened when an indicator was not defined in
the keycodes section.