Make stdin and stderr unbuffered again
freopen() always returns buffered streams, we want unbuffered ones so
that in case of a crash we still get the output.

I also added a --unsafe flag. When things segfault we still want to be
able to run gdb and crash so we know where tha problem happened. In GDB
we can execute `set follow-fork-mode child` but I'm not sure I will
remember, probably the --unsafe flag is easier to remember?. The real
solution would be to use a better debugger where it's easy to see these
problems.