Remove parse_xkb_block(), use the real parser instead
At the beginning before we had a real xkb parser, we did a simple
parsing of xkb files to split the components into those files required
by the system. Now  that we have a real parser it's better to use it
then use the writer to create these component files.

This also enforces intalling only layouts that are parseable by us.
Which at the moment, means all basic layouts from the database (no
variants or options). Maybe if we see people wanting to load other kinds
of layouts we can make a more permissive parser to split the components.