Add a prefix to installed layout components
We do this because if the system database gets an update the medatada
will be overwritten and we won't be able to detect this broken state
where we have partial layouts. Even though things would probably look
fine and from the users perspective their layouts just disappeared from
the system, it could get confusing.

We'll see how this pans out. I'm not sure what the consequences are of
changing the layout name here. Certainly people won't be able to call
'setxkbmap <layout_name>' they would have to use our prefixed name
there. But, while implementing the parser I mostly decided not caring
about support for setxkbmap this is not that big of a deal. What I don't
know, and could be an issue is how the layout will be shown to the user
in their operating system's configuration, is it going to show the
prefixed name or not? we would like to not have it prefixed. Hopefully
systems load the name from the xml metadata and not from the filenames.

I used the table idiom to centralize strings related to layout
components.