Remove the cstm_ prefix, it was a bad idea
It broke the indicator in elementary OS, then every custom layout would
be shown as cs. It also made confusing when to use the prefixed name or
where not. Instead, to allow reinstalling layouts with names used by the
default layout database, we let the user easily choose another name, and
we don't allow overwriting layouts from the default database. Custom
layouts can  be overwritten though.

To fix the problem with updates to the layout database, what we really
need is to keep track of installed custom layouts inside a file in
somewhere in the system. Every line can be a custom layout name and the
a timestamp of when it was last updated, this last field being useful
for handling the (unlikely) event where the user creates a custom layout
and then the default database adds a new layout with that same name. The
drawback is we won't allow manual modifications to the layout files...

The other issue would be where to put this file, maybe it should be
stored in /usr/share/X11/xkb/rules so it is global to the whole system.
We just need to make sure it won't be deleted when the database package
is updated.